requirements.txt
discord.py
loguru
pip
praw
pillow